updated cart service impl and cart repository to take in CartID class as ID instead of using long

+++ src/main/java/com/zhen/mypersonalshop/Repository/ProductRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface ProductRepository extends CrudRepository<Product, Integer> {
+public interface ProductRepository extends CrudRepository<Product, Long> {
 
 }
